Rename to CheckboxArrayField
