+++ src/fields/index.js
@@ -1,3 +1,3 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as CheckboxListField } from './checkboxes/src/components/CheckboxListField';
+export { default as CheckboxArrayField } from './checkboxes/src/components/CheckboxArrayField';
 export { default as DescriptionField } from './DescriptionField';
